2015-9-30: Removed vagrant, updated gulp tasks
Removed Vagrant and instead setup the styleguide to use a browsersync
server. Updated paths within the styleguide to use relative urls, and
updated the gulp tasks to pipe compiled css/js/images over to the
styleguide as an additional destination. Added an additional kss task
so that the main overview will be accessed at the / url, and sub-pages
can be accessed such as /basic and /news