Update the repo documentation
Updated the main README.md file to include more accurate instructions
based on changes to the demo site structure. Removed references to the
pages section, and instead explained how the kss-includes section
works. Updated the templates section to cover using HTML boilerplate,
or using the included templates as an example. Update the default gulp
task to run essential commands for the demo.